+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 #### Improvements
 * Fix #3316: allow locking deletion to resource version
 * Fix #3327: Removed generated ResourceHandlers
+* Fix #3349: ensuring that dsl context values always are applied over user ListOptions
 
 #### Dependency Upgrade